+++ server/src/api/tasks/tasks.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { ProjectsService } from 'src/api/projects/projects.service';
 import { Repository } from 'typeorm';
 import { CreateTaskDto } from './dto/create-task.dto';
+import { UpdateTaskDto } from './dto/update-task.dto';
 import { Task } from './tasks.entity';
 
 @Injectable()
@@ -32,11 +33,28 @@ export class TasksService {
 
   async getAllTasks() {
     const tasks = await this.taskRepository.find({ relations: ['project'] });
-    return tasks;
+    if (tasks) return tasks;
+    throw new HttpException('Задачи не найдены', HttpStatus.NOT_FOUND);
   }
 
-  async delete(id: number): Promise<Task | undefined> {
+  async delete(id: number): Promise<Task> {
     const task = await this.taskRepository.findOneBy({ id });
-    return this.taskRepository.remove(task);
+    if (task) return this.taskRepository.remove(task);
+    throw new HttpException('Задача не найдена', HttpStatus.NOT_FOUND);
+  }
+
+  async update(id: number, dto: UpdateTaskDto): Promise<Task> {
+    const task = await this.taskRepository.findOneBy({ id });
+    if (task) {
+      const newTask = await this.taskRepository.merge(task, {
+        text: dto.text,
+        responsible: dto.responsible,
+        status: dto.status,
+        isUrgently: dto.isUrgently,
+        date: dto.date,
+      });
+      return this.taskRepository.save(newTask);
+    }
+    throw new HttpException('Задача не найдена', HttpStatus.NOT_FOUND);
   }
 }
